% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zooaRch.r
\docType{data}
\name{speth83}
\alias{speth83}
\title{Bison survival data from Speth 1983}
\description{
Bison survival data from Speth 1983
}
\author{
John D. Speth
}
\references{
Speth, J. D. 1983  Bison Kills and Bone Counts:
Decision Making by Ancient Hunters. University of Chicago Press, London.
}
\keyword{datasets}

