% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooaRch.r
\docType{package}
\name{zooaRch-package}
\alias{zooaRch}
\alias{zooaRch-package}
\title{Analytical Tools for Zooarchaeological Data}
\description{
Functions in this package allow users to import, manipulate, analyze and visualize
zooarchaeological data - the faunal remains recovered from archaeological sites.
}
\author{
Erik Otarola-Castillo, Jesse Wolfhagen, and Max D. Price
}

